Reduce activation time
Reduce the time it takes to activate this package by:
* Moving package dependency checking to an idle callback
* Lazily loading the dependencies
